Enforce max-lines:500 per file; split 7 oversized files
ESLint config gains `max-lines: 500` (skipBlankLines + skipComments),
complementing the per-function `max-lines-per-function: 100`. The
per-function rule lets a 1500-line file of small functions pass lint
even though the file itself is unreadable; per-file enforcement closes
that gap.

Seven files violated and were split (raw line counts):

- App.tsx                          559 →  99  (MainPane, TerminalPane,
                                              DebugPanel, ActivityQueueLauncher,
                                              TerminalBusyOverlay extracted)
- ManageDialogView.tsx           1628 → 364   (per-tab files: General,
                                              Runtime, AI, Ports, SSH,
                                              Delete; fields + helpers)
- ActivityQueueDrawer.tsx         984 → 320   (ActivityCard +
                                              ContainerStatus + helpers)
- Sidebar.tsx                     745 →  99   (TenantGroup, EnvironmentRow,
                                              PrimaryCloudAliasControl,
                                              helpers)
- GlobalConfigDialogView.tsx      717 → 144   (CloudAliases, CloudContexts,
                                              shared, helpers)
- Titlebar.tsx                    711 →  25   (Controls, IdleStatusWidget,
                                              Status, helpers)
- manageEnvironmentThunks.ts      650 →  25   (split into manageDialogThunks,
                                              manageDeployThunks,
                                              manageDeleteThunks,
                                              manageCloudContextThunks,
                                              manageHiddenSessionThunks;
                                              original file is now a
                                              re-export facade)

All gates remain green: yarn lint (0/0), yarn typecheck, yarn
format:check, Playwright 9/9. No eslint-disable comments introduced. No
rule severities or thresholds changed beyond adding the new rule.
