+++ erun-ui/frontend/src/App.tsx
@@ -1,53 +1,25 @@
-import { CheckCircle2, ChevronDown, ChevronUp, Copy, LoaderCircle, Trash2 } from 'lucide-react';
-import { Rocket } from 'lucide-react';
 import * as React from 'react';
 
-import { useActivityQueue, useTerminalActivityLockState } from '@/app/activityQueueState';
 import { ControllerProvider } from '@/app/ControllerContext';
-import { readError } from '@/app/errors';
 import { useAppDispatch, useAppSelector } from '@/app/hooks';
-import {
-  clearDebugOutput,
-  setDebugOpen,
-  startDebugResize,
-  startReviewResize,
-  startSidebarResize,
-} from '@/app/layoutThunks';
-import { selectActiveSessionDebug } from '@/app/selectors';
+import { startSidebarResize } from '@/app/layoutThunks';
 import { setActivityQueueOpen } from '@/app/slices/layoutSlice';
-import {
-  clearHiddenLockOverlay,
-  hideLockOverlay,
-  setDebugCopyStatus,
-} from '@/app/slices/terminalStatusSlice';
 import { TerminalController } from '@/app/TerminalController';
-import { ActivityLockOverlay } from '@/components/app/ActivityLockOverlay';
-import { ActivityQueueDrawer } from '@/components/app/ActivityQueueDrawer';
+import { ActivityQueueLauncher } from '@/components/app/ActivityQueueLauncher';
 import { EnvironmentDialogView } from '@/components/app/EnvironmentDialogView';
 import { GlobalConfigDialogView } from '@/components/app/GlobalConfigDialogView';
+import { MainPane } from '@/components/app/MainPane';
 import { ManageDialogView } from '@/components/app/ManageDialogView';
 import { ReconnectDialog } from '@/components/app/ReconnectDialog';
 import { ResizeHandle } from '@/components/app/ResizeHandle';
-import { ReviewPanel } from '@/components/app/ReviewPanel';
 import { Sidebar } from '@/components/app/Sidebar';
-import { TenantDashboardView } from '@/components/app/TenantDashboardView';
 import { TenantDialogView } from '@/components/app/TenantDialogView';
-import { TerminalTabStrip } from '@/components/app/TerminalTabStrip';
 import { Titlebar } from '@/components/app/Titlebar';
-import { Button } from '@/components/ui/button';
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
+import { TooltipProvider } from '@/components/ui/tooltip';
 import { cn } from '@/lib/utils';
-
-import { ClipboardSetText } from '../wailsjs/runtime/runtime';
 
 const splitterClassName =
   'relative cursor-col-resize bg-transparent before:absolute before:top-0 before:bottom-0 before:left-1 before:w-px before:bg-transparent before:transition-colors hover:before:bg-border [.is-resizing_&]:before:bg-border';
-
-const reviewSplitterClassName =
-  'relative cursor-col-resize border-l bg-background before:absolute before:top-0 before:bottom-0 before:left-1 before:w-px before:bg-transparent before:transition-colors hover:before:bg-border [.is-resizing-review_&]:before:bg-border';
-
-const debugSplitterClassName =
-  'relative cursor-row-resize bg-[oklch(0.06_0_0)] before:absolute before:left-0 before:right-0 before:top-1 before:h-px before:bg-transparent before:transition-colors hover:before:bg-[oklch(0.36_0_0)] [.is-resizing-debug_&]:before:bg-[oklch(0.46_0_0)]';
 
 export function App(): React.ReactElement {
   const controller = React.useMemo(() => new TerminalController(), []);
@@ -124,436 +96,5 @@ export function App(): React.ReactElement {
         />
       </TooltipProvider>
     </ControllerProvider>
-  );
-}
-
-// ActivityQueueLauncher renders a floating action button that opens the
-// deploy-queue drawer. It deliberately mounts the drawer so its subscription
-// to deploy:state events is always live — even when the drawer is closed —
-// so the badge count reflects active deploys without the user having opened
-// the drawer first.
-function ActivityQueueLauncher({
-  open,
-  onOpen,
-  onClose,
-}: {
-  open: boolean;
-  onOpen: () => void;
-  onClose: () => void;
-}): React.ReactElement {
-  const { entries } = useActivityQueue();
-  const activeCount = entries.filter((entry) => entry.status === 'running').length;
-  return (
-    <>
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <Button
-            type="button"
-            variant="outline"
-            size="icon"
-            className="fixed bottom-5 right-5 z-20 size-10 rounded-full shadow-lg"
-            aria-label={
-              activeCount > 0
-                ? `Open deploy queue (${String(activeCount)} active)`
-                : 'Open deploy queue'
-            }
-            onClick={onOpen}
-          >
-            <Rocket aria-hidden="true" className="size-4" />
-            {activeCount > 0 && (
-              <span className="absolute -top-1 -right-1 inline-flex size-4 items-center justify-center rounded-full bg-blue-500 text-[10px] font-medium text-white">
-                {activeCount}
-              </span>
-            )}
-          </Button>
-        </TooltipTrigger>
-        <TooltipContent side="left">
-          {activeCount > 0
-            ? `${String(activeCount)} deploy${activeCount > 1 ? 's' : ''} in progress`
-            : 'Deploys'}
-        </TooltipContent>
-      </Tooltip>
-      <ActivityQueueDrawer open={open} onClose={onClose} />
-    </>
-  );
-}
-
-function MainPane({
-  terminalPaneRef,
-  terminalRootRef,
-  reviewViewRef,
-  reviewMainRef,
-  diffListRef,
-  onOpenActivityQueue,
-}: {
-  terminalPaneRef: React.RefObject<HTMLElement | null>;
-  terminalRootRef: React.RefObject<HTMLDivElement | null>;
-  reviewViewRef: React.RefObject<HTMLElement | null>;
-  reviewMainRef: React.RefObject<HTMLDivElement | null>;
-  diffListRef: React.RefObject<HTMLDivElement | null>;
-  onOpenActivityQueue: () => void;
-}): React.ReactElement {
-  const dashboardTenant = useAppSelector((state) => state.tenantDashboard.tenant);
-  const debugOpen = useAppSelector((state) => state.layout.debugOpen);
-  const debugOutput = useAppSelector((state) => state.terminal.debugOutput);
-  const sessionId = useAppSelector((state) => state.terminal.sessionId);
-  const verboseDebug = useAppSelector((state) => selectActiveSessionDebug(state, sessionId));
-  const dashboardOpen = Boolean(dashboardTenant);
-  return (
-    <main
-      ref={terminalPaneRef}
-      className={cn(
-        'grid h-full min-h-0 min-w-0 overflow-hidden bg-terminal',
-        dashboardOpen
-          ? 'grid-rows-[minmax(0,1fr)] bg-background'
-          : debugOpen
-            ? 'grid-rows-[minmax(0,1fr)_var(--debug-height)]'
-            : 'grid-rows-[minmax(0,1fr)_34px]',
-      )}
-    >
-      {dashboardOpen && <TenantDashboardView />}
-      <TerminalPane
-        hidden={dashboardOpen}
-        terminalRootRef={terminalRootRef}
-        reviewViewRef={reviewViewRef}
-        reviewMainRef={reviewMainRef}
-        diffListRef={diffListRef}
-        onOpenActivityQueue={onOpenActivityQueue}
-      />
-      {!dashboardOpen && (
-        <DebugPanel
-          open={debugOpen}
-          output={debugOutput}
-          sessionId={sessionId}
-          verbose={verboseDebug}
-        />
-      )}
-    </main>
-  );
-}
-
-function TerminalPane({
-  hidden,
-  terminalRootRef,
-  reviewViewRef,
-  reviewMainRef,
-  diffListRef,
-  onOpenActivityQueue,
-}: {
-  hidden: boolean;
-  terminalRootRef: React.RefObject<HTMLDivElement | null>;
-  reviewViewRef: React.RefObject<HTMLElement | null>;
-  reviewMainRef: React.RefObject<HTMLDivElement | null>;
-  diffListRef: React.RefObject<HTMLDivElement | null>;
-  onOpenActivityQueue: () => void;
-}): React.ReactElement {
-  const dispatch = useAppDispatch();
-  const sessionId = useAppSelector((state) => state.terminal.sessionId);
-  const reviewOpen = useAppSelector((state) => state.layout.reviewOpen);
-  const terminalBusy = useAppSelector((state) => state.terminalStatus.terminalBusy);
-  const terminalMessage = useAppSelector((state) => state.terminalStatus.terminalMessage);
-  const locks = useTerminalActivityLockState();
-  const hiddenForSession = useAppSelector(
-    (state) => state.terminalStatus.hiddenLockSessions[sessionId],
-  );
-  const liveLock = locks.get(sessionId) ?? null;
-  // The user can dismiss the overlay locally for a session if it's
-  // covering output they need to read or input they need to provide
-  // (e.g. the in-pod CLI's helm-recovery prompt). Backend keeps the
-  // lock state intact; only this desktop's view of it is hidden.
-  React.useEffect(() => {
-    if (!liveLock && hiddenForSession) {
-      dispatch(clearHiddenLockOverlay(sessionId));
-    }
-  }, [dispatch, liveLock, hiddenForSession, sessionId]);
-  const activeLock = liveLock && !hiddenForSession ? liveLock : null;
-  const hideActiveLock = React.useCallback(() => {
-    dispatch(hideLockOverlay(sessionId));
-  }, [dispatch, sessionId]);
-  return (
-    <div
-      className={cn(
-        'grid min-h-0 min-w-0 grid-cols-[minmax(360px,1fr)] overflow-hidden',
-        hidden && 'hidden',
-        reviewOpen &&
-          'grid-cols-[minmax(360px,1fr)_10px_minmax(420px,var(--review-width))] max-[980px]:grid-cols-[minmax(260px,1fr)_10px_minmax(360px,min(var(--review-width),58vw))]',
-      )}
-    >
-      <div className="grid h-full min-h-0 min-w-0 grid-rows-[32px_minmax(0,1fr)] overflow-hidden">
-        <TerminalTabStrip />
-        {/* Padding lives on the wrapper, not on the FitAddon parent: xterm's FitAddon reads the parent's computed height but does not subtract its padding, so any padding on terminalRoot would over-count rows and clip the bottom line. */}
-        <div
-          id="erun-terminal-pane"
-          className="relative h-full min-h-0 min-w-0 overflow-hidden box-border px-4 pt-3.5"
-        >
-          <div ref={terminalRootRef} className="terminal h-full min-h-0 min-w-0 w-full" />
-          <TerminalBusyOverlay message={terminalBusy ? terminalMessage : ''} />
-          {activeLock && (
-            <ActivityLockOverlay
-              lock={activeLock}
-              onOpenQueue={onOpenActivityQueue}
-              onProceedAnyway={hideActiveLock}
-            />
-          )}
-        </div>
-      </div>
-      <ResizeHandle
-        className={reviewSplitterClassName}
-        orientation="vertical"
-        label="Resize diff panel"
-        hidden={!reviewOpen}
-        onMouseDown={(event) => {
-          dispatch(startReviewResize(event));
-        }}
-      />
-      <ReviewPanel
-        reviewViewRef={reviewViewRef}
-        reviewMainRef={reviewMainRef}
-        diffListRef={diffListRef}
-      />
-    </div>
-  );
-}
-
-function TerminalBusyOverlay({ message }: { message: string }): React.ReactElement | null {
-  if (!message) {
-    return null;
-  }
-
-  return (
-    <div className="pointer-events-none absolute inset-0 z-10 flex items-center justify-center bg-terminal/45">
-      <div className="flex max-w-[min(520px,calc(100%-48px))] items-center gap-3 rounded-md border border-[oklch(0.28_0_0)] bg-[oklch(0.08_0_0)] px-4 py-3 text-[13px] leading-[1.35] text-[oklch(0.86_0_0)] shadow-lg">
-        <LoaderCircle
-          className="size-4 flex-none animate-spin text-[oklch(0.72_0_0)]"
-          aria-hidden="true"
-        />
-        <span className="min-w-0 truncate" title={message}>
-          {message}
-        </span>
-      </div>
-    </div>
-  );
-}
-
-function useDebugPanelScrollBinding(
-  open: boolean,
-  sessionId: number,
-  output: string,
-): {
-  outputRef: React.RefObject<HTMLDivElement | null>;
-  handleScroll: React.UIEventHandler<HTMLDivElement>;
-} {
-  const outputRef = React.useRef<HTMLDivElement>(null);
-  const stuckToBottomRef = React.useRef(true);
-
-  const handleScroll = React.useCallback(() => {
-    const el = outputRef.current;
-    if (!el) {
-      return;
-    }
-    const distanceFromBottom = el.scrollHeight - el.scrollTop - el.clientHeight;
-    stuckToBottomRef.current = distanceFromBottom <= 4;
-  }, []);
-
-  // When the active session changes, reset stick-to-bottom and snap to the
-  // bottom of the new buffer so users see the latest debug output for that tab.
-  React.useEffect(() => {
-    stuckToBottomRef.current = true;
-    if (open && outputRef.current) {
-      outputRef.current.scrollTop = outputRef.current.scrollHeight;
-    }
-  }, [open, sessionId]);
-
-  React.useEffect(() => {
-    if (!open) {
-      return;
-    }
-    const el = outputRef.current;
-    if (!el) {
-      return;
-    }
-    if (stuckToBottomRef.current) {
-      el.scrollTop = el.scrollHeight;
-    }
-  }, [open, output]);
-
-  return { outputRef, handleScroll };
-}
-
-function useDebugCopyAction(
-  output: string,
-  canCopy: boolean,
-): { copyStatus: string; copyDebugOutput: () => void } {
-  const dispatch = useAppDispatch();
-  const copyStatus = useAppSelector((state) => state.terminalStatus.debugCopyStatus);
-
-  React.useEffect(() => {
-    dispatch(setDebugCopyStatus(''));
-  }, [dispatch, output]);
-
-  const copyDebugOutput = React.useCallback(() => {
-    if (!canCopy) {
-      return;
-    }
-    void ClipboardSetText(output)
-      .then(() => {
-        dispatch(setDebugCopyStatus('Copied'));
-        window.setTimeout(() => dispatch(setDebugCopyStatus('')), 1400);
-      })
-      .catch((error: unknown) => dispatch(setDebugCopyStatus(readError(error))));
-  }, [canCopy, dispatch, output]);
-
-  return { copyStatus, copyDebugOutput };
-}
-
-function DebugPanel({
-  open,
-  output,
-  sessionId,
-  verbose,
-}: {
-  open: boolean;
-  output: string;
-  sessionId: number;
-  verbose: boolean;
-}): React.ReactElement {
-  const dispatch = useAppDispatch();
-  const canCopy = output.trim().length > 0;
-  const { outputRef, handleScroll } = useDebugPanelScrollBinding(open, sessionId, output);
-  const { copyStatus, copyDebugOutput } = useDebugCopyAction(output, canCopy);
-
-  return (
-    <section
-      className={cn(
-        'grid min-h-0 border-t border-[oklch(0.26_0_0)] bg-[oklch(0.06_0_0)] text-[oklch(0.86_0_0)]',
-        open ? 'grid-rows-[6px_34px_minmax(0,1fr)]' : 'grid-rows-[34px]',
-      )}
-    >
-      {open && (
-        <ResizeHandle
-          className={debugSplitterClassName}
-          orientation="horizontal"
-          label="Resize debug panel"
-          onMouseDown={(event) => {
-            dispatch(startDebugResize(event));
-          }}
-        />
-      )}
-      <DebugPanelHeader
-        open={open}
-        canCopy={canCopy}
-        copyStatus={copyStatus}
-        copyDebugOutput={copyDebugOutput}
-      />
-      {open && (
-        <DebugPanelOutput
-          outputRef={outputRef}
-          handleScroll={handleScroll}
-          output={output}
-          sessionId={sessionId}
-          verbose={verbose}
-        />
-      )}
-    </section>
-  );
-}
-
-function DebugPanelHeader({
-  open,
-  canCopy,
-  copyStatus,
-  copyDebugOutput,
-}: {
-  open: boolean;
-  canCopy: boolean;
-  copyStatus: string;
-  copyDebugOutput: () => void;
-}): React.ReactElement {
-  const dispatch = useAppDispatch();
-  return (
-    <div className="flex h-[34px] items-center justify-between gap-2 border-b border-[oklch(0.18_0_0)] px-3">
-      <button
-        type="button"
-        className="flex min-w-0 items-center gap-2 border-0 bg-transparent p-0 text-xs font-medium tracking-normal text-[oklch(0.76_0_0)]"
-        onClick={() => {
-          dispatch(setDebugOpen(!open));
-        }}
-      >
-        {open ? (
-          <ChevronDown className="size-4" aria-hidden="true" />
-        ) : (
-          <ChevronUp className="size-4" aria-hidden="true" />
-        )}
-        <span>Debug</span>
-        <span className="text-[11px] font-normal text-[oklch(0.56_0_0)]">
-          {open ? 'erun -vv output' : 'collapsed'}
-        </span>
-      </button>
-      {open && (
-        <div className="flex items-center gap-1">
-          <Button
-            className="h-6 px-2 text-[11px] [&_svg]:size-3.5"
-            type="button"
-            variant="ghost"
-            size="sm"
-            disabled={!canCopy}
-            onClick={copyDebugOutput}
-          >
-            {copyStatus === 'Copied' ? (
-              <CheckCircle2 aria-hidden="true" />
-            ) : (
-              <Copy aria-hidden="true" />
-            )}
-            {copyStatus || 'Copy'}
-          </Button>
-          <Button
-            className="h-6 px-2 text-[11px] [&_svg]:size-3.5"
-            type="button"
-            variant="ghost"
-            size="sm"
-            onClick={() => {
-              dispatch(clearDebugOutput());
-            }}
-          >
-            <Trash2 aria-hidden="true" />
-            Clear
-          </Button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function DebugPanelOutput({
-  outputRef,
-  handleScroll,
-  output,
-  sessionId,
-  verbose,
-}: {
-  outputRef: React.RefObject<HTMLDivElement | null>;
-  handleScroll: React.UIEventHandler<HTMLDivElement>;
-  output: string;
-  sessionId: number;
-  verbose: boolean;
-}): React.ReactElement {
-  return (
-    <div
-      ref={outputRef}
-      onScroll={handleScroll}
-      className="min-h-0 overflow-auto px-3 py-2 font-mono text-[11px] leading-[1.35] text-[oklch(0.82_0_0)]"
-    >
-      {!verbose && sessionId > 0 && (
-        <div className="mb-2 rounded border border-[oklch(0.32_0_0)] bg-[oklch(0.10_0_0)] px-3 py-2 text-[11px] leading-[1.4] text-[oklch(0.74_0_0)]">
-          Active session was not started with <code>-vv</code>, so trace lines won't appear here. To
-          capture verbose output: run <code>erun -vv &lt;command&gt;</code> in this shell, or open
-          this Debug panel before launching commands from the UI.
-        </div>
-      )}
-      <pre className="m-0 whitespace-pre-wrap break-words font-mono text-[11px] leading-[1.35]">
-        {output ||
-          'Run an environment command while Debug is expanded to stream erun -vv output here.'}
-      </pre>
-    </div>
   );
 }
