+++ erun-ui/frontend/src/app/manageDeployThunks.ts
@@ -1,0 +1,26 @@
+import { closeManageDialog } from './manageDialogThunks';
+import { selectManageRuntimeImage } from './selectors';
+import { startDeploySelection } from './sessionThunks';
+import type { AppThunk } from './store';
+import { normalizeDialogValue } from './versionSuggestions';
+
+export const submitManageDeploy = (): AppThunk<Promise<void>> => async (dispatch, getState) => {
+  const dialog = getState().manageDialog;
+  if (dialog.busy) {
+    return;
+  }
+  const selection = dialog.selection;
+  if (!selection) {
+    dispatch(closeManageDialog());
+    return;
+  }
+  const version = normalizeDialogValue(dialog.version);
+  dispatch(closeManageDialog());
+  await dispatch(
+    startDeploySelection({
+      ...selection,
+      version,
+      runtimeImage: version ? selectManageRuntimeImage(getState(), version) : '',
+    }),
+  );
+};
